requirements.txt
Flask==3.0.0
pyad==0.6.0
python-ldap==3.4.3
WTForms==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
Flask-Login==0.6.3
bcrypt==4.1.2 